+++ typescript/services/control-plane-api/src/index.ts
@@ -1,7 +1,0 @@
-export * from "./app.js";
-export * from "./auth.js";
-export * from "./migrations.js";
-export * from "./store.js";
-export * from "./types.js";
-export * from "./validation.js";
-export * from "./worker.js";